+++ com.unity.netcode.gameobjects/Tests/Runtime/RpcTypeSerializationTests.cs
@@ -14,14 +14,14 @@ using Vector4 = UnityEngine.Vector4;
 
 namespace Unity.Netcode.RuntimeTests
 {
-    public class RpcTypeSerializationTests : NetcodeIntegrationTest
+    internal class RpcTypeSerializationTests : NetcodeIntegrationTest
     {
         public RpcTypeSerializationTests()
         {
             m_UseHost = false;
         }
 
-        public class RpcTestNB : NetworkBehaviour
+        internal class RpcTestNB : NetworkBehaviour
         {
             public delegate void OnReceivedDelegate(object obj);
             public OnReceivedDelegate OnReceived;